bukti-setor-service/requirements-bukti-setor.txt
# ========================================
# REQUIREMENTS FOR BUKTI SETOR SERVICE (EASYOCR)
# ========================================

# Core Flask dependencies
Flask==2.3.3
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
SQLAlchemy==2.0.29
gunicorn==20.1.0

# Environment and utilities
python-dotenv==1.0.1
Flask-RESTful==0.3.9

# OCR and Image Processing (EasyOCR focus)
easyocr==1.7.0
opencv-python-headless==4.9.0.80
numpy==1.26.4
Pillow==10.2.0
pdf2image==1.16.3

# PyTorch (CPU only for production)
torch==2.1.0+cpu
torchvision==0.16.0+cpu
torchaudio==2.1.0+cpu
--find-links https://download.pytorch.org/whl/torch_stable.html

# Excel and Data Processing
openpyxl==3.1.2
pandas==2.2.1

# Database
psycopg2-binary==2.9.9

# Text Processing and Spell Check
pyspellchecker==0.7.2
textdistance==0.4.6
thefuzz==0.22.1

# Logging and monitoring
structlog==23.2.0
